+++ tests/unit/test_main.py
@@ -389,7 +389,7 @@ class TestMain(TestCase):
 
         response = api.get(url)
         expected_result = circuits["circuits"]
-        get_circuits.assert_called_with(archived=False)
+        get_circuits.assert_called_with(archived=False, metadata={})
         self.assertEqual(json.loads(response.data), expected_result)
 
     def test_list_with_archived_circuits_archived(self):
@@ -404,10 +404,11 @@ class TestMain(TestCase):
         get_circuits.return_value = circuits
 
         api = self.get_app_test_client(self.napp)
-        url = f"{self.server_name_url}/v2/evc/?archived=true"
+        url = f"{self.server_name_url}/v2/evc/?archived=true&metadata.a=1"
 
         response = api.get(url)
-        get_circuits.assert_called_with(archived=True)
+        get_circuits.assert_called_with(archived=True,
+                                        metadata={"metadata.a": "1"})
         expected_result = {"1": circuits["circuits"]["1"]}
         self.assertEqual(json.loads(response.data), expected_result)
 
@@ -2305,12 +2306,12 @@ class TestMain(TestCase):
             content_type="application/json",
         )
         self.assertEqual(response.status_code, 201)
-        args = self.napp.mongo_controller.update_bulk_evc.call_args[0]
+        args = self.napp.mongo_controller.update_evcs.call_args[0]
         ids = payload.pop("circuit_ids")
-        expected = {f"metadata.{k}": v for k, v in payload.items()}
         self.assertEqual(args[0], ids)
-        self.assertEqual(args[1], {"$set": expected})
-        calls = self.napp.mongo_controller.update_bulk_evc.call_count
+        self.assertEqual(args[1], payload)
+        self.assertEqual(args[2], "add")
+        calls = self.napp.mongo_controller.update_evcs.call_count
         self.assertEqual(calls, 1)
         evc_mock.extend_metadata.assert_called_with(payload)
 
@@ -2353,9 +2354,7 @@ class TestMain(TestCase):
         self.napp.circuits = {"1234": evc_mock}
         api = self.get_app_test_client(self.napp)
         payload = {
-            "circuit_ids": ["1234"],
-            "metadata1": 1,
-            "metadata2": 2
+            "circuit_ids": ["1234"]
         }
         response = api.delete(
             f"{self.server_name_url}/v2/evc/metadata/metadata1",
@@ -2363,11 +2362,10 @@ class TestMain(TestCase):
             content_type="application/json",
         )
         self.assertEqual(response.status_code, 200)
-        args = self.napp.mongo_controller.update_bulk_evc.call_args[0]
-        ids = payload.pop("circuit_ids")
-        expected = {"metadata.metadata1": ""}
-        self.assertEqual(args[0], ids)
-        self.assertEqual(args[1], {"$unset": expected})
-        calls = self.napp.mongo_controller.update_bulk_evc.call_count
+        args = self.napp.mongo_controller.update_evcs.call_args[0]
+        self.assertEqual(args[0], payload["circuit_ids"])
+        self.assertEqual(args[1], {"metadata1": ""})
+        self.assertEqual(args[2], "del")
+        calls = self.napp.mongo_controller.update_evcs.call_count
         self.assertEqual(calls, 1)
         self.assertEqual(evc_mock.remove_metadata.call_count, 1)